ddl fixed, bootstrap.cpp commented, others ((unused))

+++ src/backend/bridge/ddl/tests/bridge_test_helper.cpp
@@ -89,7 +89,7 @@ bool BridgeTest::CheckColumnWithConstraint(catalog::Column &column,
     assert(strcmp(constraint_infos[0].GetName().c_str(),
                   constraint_name.c_str()) == 0);
 
-  assert(constraint_infos.size() == constraint_count);
+  assert(constraint_infos.size() == (unsigned int)constraint_count);
   if (constraint_type == CONSTRAINT_TYPE_FOREIGN)
     assert(constraint_infos[0].GetForeignKeyListOffset() == foreign_key_offset);
 
@@ -145,8 +145,8 @@ bool BridgeTest::CheckForeignKey(catalog::ForeignKey *foreign_key,
   std::vector<std::string> pk_column_names = foreign_key->GetPKColumnNames();
   std::vector<std::string> fk_column_names = foreign_key->GetFKColumnNames();
 
-  assert(pk_column_names.size() == pk_column_names_count);
-  assert(fk_column_names.size() == fk_column_names_count);
+  assert(pk_column_names.size() == (unsigned int)pk_column_names_count);
+  assert(fk_column_names.size() == (unsigned int)fk_column_names_count);
 
   assert(foreign_key->GetUpdateAction() == fk_update_action);
   assert(foreign_key->GetDeleteAction() == fk_delete_action);
